+++ docs/Whats-New.md
@@ -416,6 +416,7 @@ New:
 - Customizing whether passengers are kicked out when an aircraft fires (by ststl)
 - Shield hit flash (by Starkku)
 - Option to scatter `Anim/SplashList` animations around impact coordinates (by Starkku)
+- Customizable rocker amplitude (by TwinkleStar)
 - Customizable wake anim (by TwinkleStar)
 
 Vanilla fixes: